requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.9
mysqlclient>=2.2.0
django-filter>=23.5
python-dotenv>=1.0.0
gunicorn>=21.2.0
django-cors-headers>=4.3.0
Pillow>=10.1.0
